one last update, not to the hash table but for the function hash table demo
I updated the demo on line 364, 365 and 366 to usekey35 and value pig to match the previous update to replace key 55 with 35. By doing that an above demo of  .put of key forces a rehash. Previously do to the resize function there was no actual demonstration of resize so I updated a few keys to show the rehash. I forget to update these 3 demo lines to now use key 35 as key 55 no longer existed. This makes all the demo work properly. There was no change to the hash table code it is good and runs good. v/r Mike