WK6: Expression Error Detection
This update, annotated with filename
buildParseTreeWED.py requires use of two modules,
the matching updated binaryTreeWED.py, and the
non updated doublyLinkedStack.py. Together the two
undated files provide a With Error Detection
capability to provide a low level expression error
detection capability for items such as redundant
parenthesis or other errors that make an expression
not proper or incorrect.

This updated buildParseTreeWED.py will
generate a helper file to the same
directory or folder these files are in called
errorDetection.txt that is a temp or buffer
that will overwrite each run. You do not need
it to start, it will create and update as
needed.
v/r Mike